requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-storage
setuptools
sqlalchemy
gunicorn
psycopg2-binary
python-jose
passlib